requirements.txt
# Core Flask & Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0

# Database & Caching
pymongo==4.5.0
redis==4.6.0
mongoengine==0.27.0

# Background Tasks & Scheduling
celery==5.3.2
celery-beat==2.4.0
schedule==1.2.0

# AI & ML APIs (Primary: Mistral, Fallback: Groq)
mistralai==0.0.12
groq==0.4.1
openai==1.3.5
anthropic==0.7.8

# Web Scraping & Content Sources
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
newsapi-python==0.2.6
praw==7.7.1

# Browser Automation (Username/Password Login)
selenium==4.15.2
playwright==1.39.0
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1

# Image & Video Processing
Pillow==10.0.1
opencv-python==4.8.1.78
moviepy==1.0.3
imageio==2.31.5
numpy==1.24.3

# Security & Encryption
cryptography==41.0.7
bcrypt==4.0.1
passlib==1.7.4

# Social Media APIs (Official when available)
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.103.0
facebook-sdk==3.1.0
tweepy==4.14.0
python-linkedin-v2==2.0.1

# Data Processing & Utils
marshmallow==3.20.1
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.3
validators==0.22.0

# Monitoring & Logging
flask-limiter==3.5.0
structlog==23.1.0

# Testing (Optional)
pytest==7.4.3
pytest-flask==1.3.0